Add checks for existing files

+++ install-server.sh
@@ -6,16 +6,14 @@
 INSTALL_DIRECTORY=/etc/wireguard
 SERVER_PRIVATE=server_private_key
 SERVER_PUBLIC=server_public_key
-OVERWRITE=0
 
-# Set IP range (experimental)
+# Set IP range if none specified (experimental)
 if [ $# -eq 0 ]
 then
 	server_ip="10.100.200.1"
 else
 	server_ip=$1
 fi
-
 
 # Ubuntu
 sudo apt-get update
@@ -30,26 +28,41 @@ sudo apt-get install -y qrencode
 
 # Create Server Keys
 sudo -i
-if [[ -d "$INSTALL_DIRECTORY" ]]
+if [ -d "$INSTALL_DIRECTORY" ]
 then
 	echo "$INSTALL_DIRECTORY exists"
 	echo "This process could over-write existing keys!"
+	echo
+	while true; do
+		read -p "Do you wish to overwrite existing keys?" yn
+		case $yn in
+			[Yy]* ) OVERWRITE=1;;
+			[Nn]* ) OVERWRITE=0;;
+			* ) echo "Please answer yes or no.";;
+		esac
+	done
 else
 	echo "Creating $INSTALL_DIRECTORY"
 	mkdir -m 0700 $INSTALL_DIRECTORY
 fi
 
 cd $INSTALL_DIRECTORY
-umask 077; wg genkey | tee $SERVER_PRIVATE | wg pubkey > $SERVER_PUBLIC
+
+if [ -f $SERVER_PRIVATE ] && [ $OVERWRITE == 0 ]
+then
+	echo "$SERVER_PRIVATE exists, skipping."
+else
+	umask 077; wg genkey | tee $SERVER_PRIVATE | wg pubkey > $SERVER_PUBLIC
+fi
 
 # Get config
 sudo wget https://raw.githubusercontent.com/rdbh/wireguard-scripts/master/wg0-server.example.conf 
 sudo wget https://raw.githubusercontent.com/rdbh/wireguard-scripts/master/wg0-client.example.conf
 
 # Check if wg0.conf already exists
-if [[ -f $INSTALL_DIRECTORY/wg0.conf ]]
+if [ -f $INSTALL_DIRECTORY/wg0.conf ] && [ $OVERWRITE == 0 ] 
 then
-	echo "$INSTALL_DIRECTORY/wg0.conf exists"
+	echo "$INSTALL_DIRECTORY/wg0.conf exists, skipping."
 else
 	# Add server key to config
 	SERVER_PUB_KEY=$(cat $INSTALL_DIRECTORY/$SERVER_PUBLIC)
@@ -79,5 +92,5 @@ sudo ufw allow 41194/udp
 
 # Use this to forward traffic from the server
 #sed -i 's/#net.ipv4.ip_forward=1/net.ipv4.ip_forward=1/g' /etc/sysctl.conf
-
+#sudo sysctl -p /etc/sysctl.conf
